devices_long command, fix regex parsing for offline devices
When an adb device is offline, the `devices-ls` command output does not contain
product, model, or device attributes (besides the state being
`offline` of course).
e.g.
<IDENTIFIER>            offline usb:1-10 transport_id:14911
<IDENTIFIER>            device usb:1-9 product:<product> model:<MODEL> device:<device> transport_id:14703

In order to support both formats, the regex has changed to use
named groups, with the group of all these three being optional. When
creating the DeviceLong struct, if these fields are unknown they are
set to "Unk".